+++ Route.cpp
@@ -51,7 +51,7 @@ void Route::setDestination(Airport* destinationIn)
 	this->destination = destinationIn;
 }
 
-void Route::setDistance(int distanceIn)
+void Route::setDistance(double distanceIn)
 {
 	this->distance = distanceIn;
 }
@@ -73,7 +73,7 @@ Airport* Route::getDestination()
 	return this->destination;
 }
 
-int Route::getDistance()
+double Route::getDistance()
 {
 	return this->distance;
 }
@@ -82,7 +82,7 @@ int Route::getDistance()
 *                    OTHER FUNCTIONS
 ********************************************************/
 
-int calculateRouteDistance(Airport*, Airport*)
+double calculateRouteDistance(Airport*, Airport*)
 {
-	float testPi = M_PI;
+	double testPi = M_PI;
 }
